+++ resources/views/pages/invoices/invoice_services_details.blade.php
@@ -264,7 +264,7 @@
                     </span>
                     @if ($invoice->isPaid == 'لم يتم الدفع' && $invoice->customer->contract)
                         @php
-                            $paymentDueDate = Carbon\Carbon::parse($invoice->date)->addDays($invoice->customer->contract->payment_grace_period ?? 0);
+                            $paymentDueDate = Carbon\Carbon::parse($invoice->date)->addDays((int) ($invoice->customer->contract->payment_grace_period ?? 0));
                             $lateDays = Carbon\Carbon::now()->gt($paymentDueDate) ? Carbon\Carbon::parse($paymentDueDate)->diffInDays(Carbon\Carbon::now()) : 0;
                         @endphp
                         <span class="badge bg-primary fs-6 px-3 py-2">
